/producer routes have a separate file now from server.js file
Clean up JS and address minor inconsistencies, including Minor CSS tweaks
