+++ server/handlers/ProducerHandler.js
@@ -1,6 +1,7 @@
 //* Update the user data on the playlist after detecting the client-side fetch request through server.js
 const User = require('../models/User');
 
+
 const updateCurrPlaylist = (userID, listName, newData) => {
     return new Promise((resolve, reject) => {
         User.findById(userID)
@@ -25,6 +26,41 @@ const updateCurrPlaylist = (userID, listName, newData) => {
     });
 }
 
+
+const renameCurrPlaylist = (userID, oldName, newName) => {
+	return new Promise((resolve, reject) => {
+		User.findById(userID)
+			.then(user => {
+				if (!user) {
+					// console.error("User not found");
+					return reject(new Error("User not found"));
+				}
+				// update playlist names list
+				const index = user.playlists.names.indexOf(oldName);
+				user.playlists.names[index] = newName;
+				// copy data from old property name
+				user.playlists.data[newName] = user.playlists.data[oldName];
+
+				// delete the old property
+				delete user.playlists.data[oldName];
+				user.currPlaylist = newName;
+
+				user.markModified(`playlists`);
+				user.markModified(`playlists.names`);
+				user.markModified(`playlists.data.${newName}`); // by default MongoDB does not check sub-properties
+				return user.save();
+			})
+			.then(() => {
+				resolve("Playlist renamed successfully"); // Note: this message will not show up in terminal
+			})
+			.catch(err => {
+				// console.error("Error updating playlists", err);
+				reject(new Error("Error renaming playlist"));
+			});
+	});
+}
+
+
 const createPlaylist = (userID, listName) => {
 	return new Promise((resolve, reject) => {
 		User.findById(userID)
@@ -38,8 +74,8 @@ const createPlaylist = (userID, listName) => {
 					new Error("Playlist already exists"); return;
 				}
 
-				
 				user.playlists.data[listName] = []; // initializes empty object for that playlist
+
 				user.markModified(`playlists.data.${listName}`);
 				user.markModified(`playlists.names`);
 				return user.save();
@@ -53,6 +89,7 @@ const createPlaylist = (userID, listName) => {
 	})
 }
 
+
 const deletePlaylist = (userID, listName) => {
 	return new Promise((resolve, reject) => {
 		User.findById(userID)
@@ -61,7 +98,7 @@ const deletePlaylist = (userID, listName) => {
 
 				// remove playlist
 				user.playlists.names = user.playlists.names.filter(name => name != listName);
-				// console.log(`Deleting: `, listName);
+
 				delete user.playlists.data[`${listName}`];
 				
 				user.markModified(`playlists`);
@@ -78,11 +115,10 @@ const deletePlaylist = (userID, listName) => {
 	});
 }
 
-const userHandler = require('./userHandler.js');
-
 
 module.exports = {
     updateCurrPlaylist,
+	renameCurrPlaylist,
 	createPlaylist,
 	deletePlaylist
 }
